feat(backend): add openfeign test between module commodity and module shop

+++ backend/mall-commodity/src/main/java/org/frogsoft/mall/commodity/controller/request/AddProductRequset.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class AddProductRequset implements Serializable {
 
-    private int shopId;
+    private Long shopId;
     private String category;
     private String brand;
     private String productName;
